# logicalplan: add max aggregate function