This commit adds an aggregate function to find the maximum value of an int64
column. This will be specifically useful to find the latest timestamp to query
over a range in a historical dataset in benchmarks.