@asubiotto 2022-08-22
I still need to add tests to `aggregate_test.go` but wanted to see if I was on the right track + get a couple of questions answered (marked as TODO in the code). cc @brancz @thorfour @metalmatze