@metalmatze 2022-08-22
Looks like a good approach. :+1: