@asubiotto 2022-08-23
Thanks for the great suggestion! I'll run some benchmarks.