@asubiotto 2022-08-24
Looks like using a concrete type is nearly twice as fast:
```
$ benchstat generic int64
name       old time/op  new time/op  delta
FindMax-8   657ns ± 0%   334ns ± 0%   ~     (p=1.000 n=1+1)
```
Will add a comment