@asubiotto 2022-08-24
Added the tests. This is basically ready for a merge to unblock #166 if the tests look good.